+++ src/sort_roms/test_order_prefer_flags.rs
@@ -5,33 +5,33 @@ async fn test() {
     // given
     let game_a = Game {
         id: 1,
-        name: String::from("Game (USA) (Rev 1)"),
+        name: String::from("Game (USA) (Rumble Version)"),
         description: String::from(""),
         comment: None,
         external_id: None,
         bios: false,
         jbfolder: false,
-        regions: String::from(""),
+        regions: String::from("US"),
         sorting: Sorting::AllRegions as i64,
         complete: true,
         system_id: 1,
-        parent_id: Some(3),
+        parent_id: Some(1),
         bios_id: None,
         playlist_id: None,
     };
     let game_b = Game {
         id: 1,
-        name: String::from("Game (USA) (Rev 2)"),
+        name: String::from("Game (USA)"),
         description: String::from(""),
         comment: None,
         external_id: None,
         bios: false,
         jbfolder: false,
-        regions: String::from(""),
+        regions: String::from("US"),
         sorting: Sorting::AllRegions as i64,
         complete: true,
         system_id: 1,
-        parent_id: Some(3),
+        parent_id: Some(1),
         bios_id: None,
         playlist_id: None,
     };
@@ -41,11 +41,11 @@ async fn test() {
         &game_a,
         &game_b,
         false,
-        &PreferRegion::None,
-        &PreferVersion::New,
-        &[],
+        &PreferredRegion::None,
+        &PreferredVersion::None,
+        &["Rumble Version"],
     );
 
     // then
-    assert_eq!(ordering, Ordering::Greater);
+    assert_eq!(ordering, Ordering::Less);
 }