+++ src/sort_roms/test_sort_1g1r_catch_all.rs
@@ -54,15 +54,15 @@ async fn test() {
         .await
         .unwrap();
 
-    let matches = subcommand().get_matches_from(&["config", "-y"]);
     let all_regions = vec![Region::Unknown];
     let one_regions = vec![Region::UnitedStates, Region::Europe];
 
     // when
     sort_system(
         &mut connection,
-        &matches,
         &progress_bar,
+        true,
+        false,
         &system,
         &all_regions,
         &one_regions,
@@ -70,8 +70,8 @@ async fn test() {
         &[],
         &[],
         true,
-        &PreferRegion::None,
-        &PreferVersion::None,
+        &PreferredRegion::None,
+        &PreferredVersion::None,
         &[],
         &SubfolderScheme::None,
         &SubfolderScheme::None,