+++ src/sort_roms/test_sort_1g1r_playlist.rs
@@ -1,9 +1,12 @@
+use super::super::config::{PreferredRegion, PreferredVersion};
 use super::super::database::*;
+use super::super::generate_playlists;
 use super::super::import_dats;
 use super::super::import_roms;
-use super::super::util::*;
 use super::*;
 use async_std::fs;
+use async_std::path::Path;
+use std::env;
 use tempfile::{NamedTempFile, TempDir};
 
 #[async_std::test]
@@ -12,6 +15,14 @@ async fn test() {
     let _guard = MUTEX.lock().await;
 
     let test_directory = Path::new("tests");
+    env::set_var(
+        "PATH",
+        format!(
+            "{}:{}",
+            test_directory.as_os_str().to_str().unwrap(),
+            env::var("PATH").unwrap()
+        ),
+    );
     let progress_bar = ProgressBar::hidden();
 
     let db_file = NamedTempFile::new().unwrap();
@@ -23,21 +34,27 @@ async fn test() {
     let tmp_directory = TempDir::new_in(&test_directory).unwrap();
     let tmp_directory = set_tmp_directory(PathBuf::from(tmp_directory.path()));
 
-    let matches = import_dats::subcommand()
-        .get_matches_from(&["import-dats", "tests/Test System (20200721).dat"]);
+    let matches = import_dats::subcommand().get_matches_from(&[
+        "import-dats",
+        "tests/Test System (20230105) (Multiple Discs).dat",
+    ]);
     import_dats::main(&mut connection, &matches, &progress_bar)
         .await
         .unwrap();
 
-    let romfile_names = vec![
-        "Test Game (Asia).rom",
-        "Test Game (Japan).rom",
-        "Test Game (USA, Europe).rom",
-        "Test Game (USA, Europe) (Beta).rom",
-    ];
+    let system = find_systems(&mut connection).await.remove(0);
+    let system_directory = get_system_directory(&mut connection, &progress_bar, &system)
+        .await
+        .unwrap();
+
+    let mut romfile_names: Vec<String> = Vec::new();
+    for i in 1..=2 {
+        romfile_names.push(format!("Test Game (USA, Europe) (Disc {}).iso", i));
+    }
+
     for romfile_name in &romfile_names {
-        let romfile_path = tmp_directory.join(romfile_name);
-        fs::copy(test_directory.join(romfile_name), &romfile_path)
+        let romfile_path = tmp_directory.join(&romfile_name);
+        fs::copy(test_directory.join(&romfile_name), &romfile_path)
             .await
             .unwrap();
         let matches = import_roms::subcommand()
@@ -47,8 +64,8 @@ async fn test() {
             .unwrap();
     }
 
-    let system = find_systems(&mut connection).await.remove(0);
-    let system_directory = get_system_directory(&mut connection, &progress_bar, &system)
+    let matches = generate_playlists::subcommand().get_matches_from(&["generate-playlists", "-a"]);
+    generate_playlists::main(&mut connection, &matches, &progress_bar)
         .await
         .unwrap();
 
@@ -79,18 +96,17 @@ async fn test() {
     .unwrap();
 
     // then
-    let romfiles = find_romfiles_by_system_id(&mut connection, system.id).await;
-    assert_eq!(4, romfiles.len());
+    let romfiles = find_romfiles(&mut connection).await;
+    assert_eq!(romfiles.len(), 3);
 
-    let one_regions_indices = vec![2, 3];
-    let trash_indices = vec![0, 1];
+    let roms_indices = vec![0, 1];
 
-    for i in one_regions_indices {
+    for i in roms_indices {
         let romfile = romfiles.get(i).unwrap();
         assert_eq!(
             &system_directory
                 .join("1G1R")
-                .join(&romfile_names.get(i).unwrap())
+                .join(romfile_names.get(i).unwrap())
                 .as_os_str()
                 .to_str()
                 .unwrap(),
@@ -99,17 +115,15 @@ async fn test() {
         assert!(Path::new(&romfile.path).is_file().await);
     }
 
-    for i in trash_indices {
-        let romfile = romfiles.get(i).unwrap();
-        assert_eq!(
-            &system_directory
-                .join("Trash")
-                .join(&romfile_names.get(i).unwrap())
-                .as_os_str()
-                .to_str()
-                .unwrap(),
-            &romfile.path
-        );
-        assert!(Path::new(&romfile.path).is_file().await);
-    }
+    let romfile = romfiles.get(2).unwrap();
+    assert_eq!(
+        &system_directory
+            .join("1G1R")
+            .join("Test Game (USA, Europe).m3u")
+            .as_os_str()
+            .to_str()
+            .unwrap(),
+        &romfile.path
+    );
+    assert!(Path::new(&romfile.path).is_file().await);
 }